MagicGradients.Tests/Parser/CssLinearGradientParserTests.cs
using FluentAssertions;
using MagicGradients.Parser;
using System.Linq;
using Xunit;

namespace MagicGradients.Tests.Parser
{
    public class CssLinearGradientParserTests
    {
        [Theory]
        [MemberData(nameof(CssLinearGradientParserTestData.SimpleGradientData), MemberType = typeof(CssLinearGradientParserTestData))]
        public void ParseCss_SimpleGradientData_CorrectlyParsed(string css, LinearGradient expected)
        {
            var parser = new CssLinearGradientParser();
            var gradient = parser.ParseCss(css).First();

            gradient.Should().BeEquivalentTo(expected);
        }
    }
}
